+++ gensim/models/wrappers/__init__.py
@@ -5,3 +5,4 @@ This package contains wrappers for other topic modeling programs.
 from .ldamallet import LdaMallet
 from .dtmmodel import DtmModel
 from .ldavowpalwabbit import LdaVowpalWabbit
+from .fasttext import FastText